+++ src/utils.ts
@@ -157,12 +157,8 @@ export function getGitLabBranchNameFromIssueNumberAndTitleAndTaskId(
 
 export async function checkWorkingTreeClean() {
   const result = await promiseSpawn("git", ["status"], "pipe");
-  if (
-    !result.stdout.includes("Your branch is up to date with") ||
-    !result.stdout.includes("nothing to commit, working tree clean")
-  ) {
-    throw Error(
-      "Working tree is not clean or something is not pushed. Aborted."
-    );
-  }
+  return (
+    result.stdout.includes("Your branch is up to date with") &&
+    result.stdout.includes("nothing to commit, working tree clean")
+  );
 }